Slight xss ish fix

+++ api/api-delete-own-user.php
@@ -1,7 +1,9 @@
 <?php
-header('Content-Type: application/json');
+ob_start();
+session_start();
 require_once __DIR__.'/../_.php';
 
+ob_end_flush();
 try {
     $data = json_decode(file_get_contents('php://input'), true);
 
@@ -11,12 +13,21 @@ try {
 
     $user_id = $data['user_id'];
 
-    $db = _db();
-    $q = $db->prepare('DELETE FROM users WHERE user_id = :user_id');
-    $q->bindParam(':user_id', $user_id);
-    $q->execute();
+    //Extra security to prevent a user from deleting another user by changing html inspect values
+    if((int)$_SESSION['user']['user_id'] === (int)$user_id){
+        var_dump($user_id);
+        var_dump($_SESSION['user']['user_id']);
 
-    echo json_encode(["info" => "User Deleted by user [id: $user_id]"]);
+        $db = _db();
+        $q = $db->prepare('DELETE FROM users WHERE user_id = :user_id');
+        $q->bindParam(':user_id', $user_id);
+        $q->execute();
+
+        echo json_encode(["info" => "User Deleted by user [id: $user_id]"]);
+    } else {
+        echo json_encode(["info" => "User ID did not match Session ID"]);
+    }
+
 } catch (Exception $e) {
     $status_code = !ctype_digit($e->getCode()) ? 500 : $e->getCode();
     $message = strlen($e->getMessage()) == 0 ? 'error - '.$e->getLine() : $e->getMessage();
